Extract common code into an abstract class
